+++ core/dbt/clients/jinja.py
@@ -21,9 +21,8 @@ from dbt.utils import (
     get_docs_macro_name,
     get_materialization_macro_name,
     get_test_macro_name,
-    deep_map_render,
 )
-
+from dbt.common.utils import deep_map_render
 from dbt.clients._jinja_blocks import BlockIterator, BlockData, BlockTag
 from dbt.contracts.graph.nodes import GenericTestNode
