+++ core/dbt/common/utils/__init__.py
@@ -1,0 +1,21 @@
+from dbt.common.utils.encoding import (
+    md5,
+    JSONEncoder,
+)
+
+from dbt.common.utils.casting import (
+    cast_to_str,
+    cast_to_int,
+    cast_dict_to_dict_of_strings,
+)
+
+from dbt.common.utils.dict import (
+    AttrDict,
+    filter_null_values,
+    merge,
+    deep_merge,
+    deep_merge_item,
+    deep_map_render,
+)
+
+from dbt.common.utils.executor import executor
