+++ core/dbt/contracts/relation.py
@@ -10,7 +10,7 @@ from dbt.common.dataclass_schema import dbtClassMixin, StrEnum
 
 from dbt.contracts.util import Replaceable
 from dbt.exceptions import CompilationError, DataclassNotDictError
-from dbt.utils import deep_merge
+from dbt.common.utils import deep_merge
 
 
 class RelationType(StrEnum):
