+++ core/dbt/contracts/results.py
@@ -15,7 +15,8 @@ from dbt.common.events.types import TimingInfoCollected
 from dbt.common.events.contextvars import get_node_info
 from dbt.common.events.helpers import datetime_to_json_string
 from dbt.logger import TimingProcessor
-from dbt.utils import lowercase, cast_to_str, cast_to_int
+from dbt.utils import lowercase
+from dbt.common.utils import cast_to_str, cast_to_int
 from dbt.common.dataclass_schema import dbtClassMixin, StrEnum
 
 import agate
